simplified serializers and views so that PCR data is always included
