+++ courses/urls.py
@@ -4,7 +4,6 @@ from . import views
 
 
 urlpatterns = [
-    path('<slug:semester>/sections/', views.SectionList.as_view()),
     path('<slug:semester>/courses/', views.CourseList.as_view()),
     path('<slug:semester>/courses/<slug:full_code>/',  views.CourseDetail.as_view()),
     path('<slug:semester>/requirements/', views.RequirementList.as_view()),
