+++ plan/views.py
@@ -1,15 +1,13 @@
 from django.db.models import Prefetch
 
-from courses.views import CourseDetail, CourseList
+from courses.views import CourseList
 from plan.filters import bound_filter, requirement_filter
 from plan.search import TypedSearchBackend
-from plan.serializers import CourseDetailWithReviewSerializer, CourseListWithReviewSerializer
 
 
 class CourseListSearch(CourseList):
     filter_backends = [TypedSearchBackend]
     search_fields = ('full_code', 'title', 'sections__instructors__name')
-    serializer_class = CourseListWithReviewSerializer
 
     def get_queryset(self):
         queryset = super().get_queryset().prefetch_related(Prefetch('sections'))
@@ -28,7 +26,3 @@ class CourseListSearch(CourseList):
                 queryset = filter_func(queryset, param, self.get_semester())
 
         return queryset.distinct()
-
-
-class CourseDetailSearch(CourseDetail):
-    serializer_class = CourseDetailWithReviewSerializer
